Add files via upload
Background:

In computers, memory is located in different parts of the computer; memory on disk is far away from the CPU and takes a long time to access, but offers a large amount of storage; cache memory is close to the CPU, so it is faster to access, but does not offer a lot of storage space.

It is therefore wise to use the cache memory in a smart way, so that pages of memory that are more likely to be accessed are already in the cache. This means that there are fewer accesses to disk memory, and the memory transaction is ultimately quicker.

In this assignment, you will use Python to simulate a cache as explained below.

The cache memory operates in the following manner:

Pages from memory are requested. When a page from memory is requested, the cache memory is searched for that page. 

If the page is found in the cache memory, then this is called a "hit".
If the page is not found in the cache memory, then this is called a "miss".  When a miss happens, the page must be retrieved from the main memory (the disk memory) and must be placed in the cache memory. If the cache memory is not full, then the page is simply added to the cache memory. If the cache memory is full, then one of the pages that is in the cache memory will have to be replaced. In this case, we say that a new page is added to the cache memory, and the old page is evicted. 
In this lab we will implement the First in First Out (FIFO) cache memory, the page that is evicted is the one that has the longest time since it was added.

Specifics:

In this exercise, we will assume that

every page is represented by a positive integer.
the capacity of the cache memory is 8 pages. 
In particular, a request for a page will be indicated by a number, e.g., the number 3 means that we request page 3. If the requested page is in the cache memory, then this will be a hit. If the requested page is not in the cache memory, then this will be a miss. In this latter case, the page has to be retrieved from the main memory and placed into the cache memory. If the cache memory is not full (i.e., it has fewer than 8 pages already in it), then we can simply add the requested page to the cache memory. If the cache memory is full (meaning that it has exactly 8 pages in in), then we have to evict one page already in the cache memory, in order to bring the newly requested page in. Which page to evict is decided based on the FIFO algorithm.

 

Program Structure
Your task is to write a program that implements the memory management operation, as described above.

Constructing the list of requests: Your program should ask the user for an integer repeatedly until 0 is entered. As mentioned above, these integers represent requested pages, and 0 signifies the end of the request input. The inputted integers should be placed into a list called requests.

If the user chooses Q, then the program should terminate. The tests will only enter Q as the first input in a batch.

Once 0 is seen, the program should iterate over the requests in the list requests, and “request” each “page” (represented simply as an integer) from the cache memory.

If the “page” is already in the cache memory cache, then this is a hit and the word “hit” should be printed to the screen.
If the “page” is not in the cache memory cache, then this is a miss and the word “miss” should be printed to screen.

In the event of a miss, the correct “page” should be evicted from the cache memory and the newly requested “page” should be correctly inserted into cache memory, according to the eviction algorithm that has been selected by the user earlier. In reality, this means that your program will remove an integer from cache and insert a new integer into cache.

Important: The newly introduced page should be added to the end of the cache list. For example, if we have the cache [4, 6, 11, 77, 32, 5, 7, 8] and a new page to add, e.g., page 20, we evict some page according to our algorithm (page 4) and the cache now becomes [6, 11, 77, 32, 5, 7, 8, 20]. If your list appears in a different order you might not pass some of the tests.
At the end of processing all requests, you should then print the final state of cache (i.e., the contents of the list cache) and the user is prompted to enter a new batch of requests. Both the request and the cache should be cleared before reading the next batch.