Create reading1.py
Open two files, dictionary.txt and text.txt.  Each line of the dictionary file contains a single word (a sequence of characters that do not include a space). We want to compute the number of times each word in the dictionary appears in the text.txt file. The text file has many lines and many words on each line.  A word should match exactly when separated by spaces, but is case-insensitive, so if hi is in the dictionary it does not appear in this is a hill because hill is not the same as hi. However, hi is the same as Hi.  Every word in the dictionary will be in lower case.

Print, one per line, in the same order as the dictionary file word: count, where count is an integer of how many times it appears in the dictionary.